backend/requirements.txt
# Web框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Strands Agent框架
strands-agents>=0.1.0

# 数据库
sqlalchemy>=2.0.35
alembic>=1.13.0

# AWS SDK
boto3>=1.35.0
opensearch-py>=2.7.0
requests-aws4auth>=1.2.0

# PDF处理 (需要GPU支持)
marker-pdf>=1.10.0

# 文本处理
langchain>=0.3.0
langchain-text-splitters>=0.3.0
tiktoken>=0.7.0

# 数据验证
pydantic>=2.9.0
pydantic-settings>=2.5.0

# 日志
structlog>=24.4.0

# 工具库
python-dotenv>=1.0.0
httpx>=0.27.0

# 开发依赖
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.8.0
isort>=5.13.0
mypy>=1.11.0
